+++ R/utils.R
@@ -78,13 +78,23 @@ get_schema.src_postgres <- function(con, schema_name, pkg, ...) {
 #' Download only those files that don't already exist
 #' @param obj an \code{\link{etl}} object
 #' @param src a character vector of URLs that you want to download
+#' @param new_filenames an optional character vector of filenames for the new (local) files. Defaults to
+#' having the same filenames as those in \code{src}.
 #' @param ... arguments passed to \code{\link[utils]{download.file}}
 #' @details Downloads only those files in \code{src} that are not already present in
 #' the directory specified by the \code{raw_dir} attribute of \code{obj}.
 #' @author idiom courtesy of Hadley Wickham
 #' @export
-smart_download <- function(obj, src, ...) {
-  lcl <- paste0(attr(obj, "raw_dir"), "/", basename(src))
+#'
+#' @examples
+#' cars <- etl("mtcars")
+#' urls <- c("http://www.google.com", "http://www.nytimes.com")
+#' smart_download(cars, src = urls)
+smart_download <- function(obj, src, new_filenames = basename(src), ...) {
+  if (length(src) != length(new_filenames)) {
+    stop("src and new_filenames must be of the same length")
+  }
+  lcl <- paste0(attr(obj, "raw_dir"), "/", new_filenames)
   missing <- !file.exists(lcl)
   mapply(utils::download.file, src[missing], lcl[missing], ... = ...)
 }
@@ -143,3 +153,49 @@ valid_year_month <- function(years, months, begin = "1970-01-01", end = Sys.Date
 
 
 
+#' Match year and month vectors to filenames
+#' @description Match year and month vectors to filenames
+#' @inheritParams extract_date_from_filename
+#' @param years a numeric vector of years
+#' @param months a numberic vector of months
+#' @return a character vector of \code{files} that match the \code{pattern}, \code{year}, and \code{month} arguments
+#' @importFrom lubridate year month
+#' @export
+#' @examples
+#' \dontrun{
+#' if (require(airlines)) {
+#'   airlines <- etl("airlines", dir = "~/dumps/airlines") %>%
+#'     etl_extract(year = 1987)
+#'   summary(airlines)
+#'   match_files_by_year_months(list.files(attr(airlines, "raw_dir")),
+#'     pattern = "On_Time_On_Time_Performance_%Y_%m.zip"), year = 1987)
+#' }
+#' }
+
+match_files_by_year_months <- function(files, pattern, years = as.numeric(format(Sys.Date(), '%Y')), months = 1:12, ...) {
+  file_df <- data.frame(filename = files, file_date = extract_date_from_filename(files, pattern)) %>%
+    mutate_(file_year = ~lubridate::year(file_date),
+            file_month = ~lubridate::month(file_date))
+  valid <- valid_year_month(years, months)
+  good <- file_df %>%
+    left_join(valid, by = c("file_year" = "year", "file_month" = "month")) %>%
+    filter_(~!is.na(month_begin))
+  return(as.character(good$filename))
+}
+
+#' @description Extracts a date from filenames
+#' @param files a character vector of filenames
+#' @param pattern a regular expression to be passed to \code{\link[lubridate]{fast_strptime}}
+#' @param ... arguments passed to \code{\link[lubridate]{fast_strptime}}
+#' @return a vector of \code{\link{POSIXct}} dates matching the pattern
+#' @importFrom lubridate fast_strptime days
+#' @export
+#' @rdname match_files_by_year_months
+
+extract_date_from_filename <- function(files, pattern, ...) {
+  files %>%
+    basename() %>%
+    lubridate::fast_strptime(format = pattern, ...) %>%
+    # why does it always return the previous day?
+    as.Date() + lubridate::days(1)
+}