+++ tests/testthat/test-etl.R
@@ -28,3 +28,16 @@ test_that("dplyr works", {
     collect()
   expect_equal(nrow(res2), 2 * nrow(mtcars))
 })
+
+
+test_that("mysql works", {
+  db <- src_mysql(default.file = "~/.my.cnf", dbname = "mtcars",
+                  user = NULL, password = NULL)
+  cars <- etl("mtcars", db = db)
+  class(cars)
+  cars %>% etl_create()
+})
+
+test_that("valid_year_month works", {
+  expect_equal(nrow(valid_year_month(years = 1999:2001, months = c(1:3, 7))), 12)
+})